backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
SpeechRecognition
google-generativeai
python-dotenv
pyaudio
openai
pydantic-settings
python-dotenv
flake8
pytest